LayoutTests/fast/css/variables/color-hex-test-expected.txt

CONSOLE MESSAGE: line 27: TypeError: Result of expression 'declaration' [undefined] is not an object.
FAILED
